requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==5.1.3
django-appconf==1.0.6
django-bootstrap5==24.3
django-imagekit==5.0.0
django-pictures==1.3.3
gunicorn==23.0.0
mysqlclient==2.2.5
packaging==24.2
pilkit==3.0
pillow==11.0.0
PyMySQL==1.1.1
soupsieve==2.6
sqlparse==0.5.1
whitenoise==6.8.2
